+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 ## CHANGELOG
 
----
-
 ### 2.11.2
-
+- [31626 Tracking de eventos Algolia](https://dev.azure.com/ARECO/AReCO/_workitems/edit/31626)
 - [30966 - Informação troca e devolução PDP [app]](https://dev.azure.com/ARECO/AReCO/_workitems/edit/30966/)
 
 ---